requirements.txt
# Core dependencies
python-dotenv>=1.0.0
python-telegram-bot==20.7
requests>=2.31.0

# Development and testing
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0